requirements.txt
Django==1.8.5
django-admin-bootstrapped==2.5.6
django-braces==1.8.1
django-cors-headers==1.1.0
django-enumfield==1.2.1
django-oauth-toolkit==0.7.2
djangorestframework==3.3.2
django-sortedm2m==1.3.0
django-push-notifications==1.4.1
httplib2==0.9.2
lazy-object-proxy==1.2.1
oauthlib==1.0.3
psycopg2==2.6.1
pylint==1.5.3
requests==2.9.1
six==1.10.0
wheel==0.24.0
wrapt==1.10.6
astroid==1.4.5
colorama==0.3.7
django-sortedm2m==1.3.0
djoser==0.4.3
ExifRead==2.1.2
Pillow==3.2.0
pip==8.1.2
python-dateutil==2.5.3
setuptools==21.2.2
rlextra==3.3.12
reportlab==3.3.12
pyRXP==2.1.1
preppy==2.3.5
Pmw==2.0.0
